Map out all characters on the license plate
Using CCA again, we can identify each character on the license plate.
Once identified, each character is resize to 20px by 20px. This is done
for the stage of character recognition. To keep track of the character
order, the column_list variable takes note of the starting x-axis of
each region. This can then be sorted to know the correct order of the
characters.

The mapping of the characters on the license plate is using the license
plate identified by me from the different regions cca.py gave. Later, I
will tweak this code so the computer can do it itself.